Cleaned up framework
