+++ Shared/Model/Extensions/UIColor+Extensions.swift
@@ -1,0 +1,8 @@
+//
+//  UIColor+Extensions.swift
+//  Sprout
+//
+//  Created by Ryan Thally on 6/23/21.
+//
+
+import Foundation
