+++ SproutKit/Sources/SproutKit/Core Data/FetchRequests/SproutPlantMO+FetchRequests.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by Ryan Thally on 6/25/21.
+//
+
+import Foundation
