+++ SproutKit/Sources/SproutKit/Models/SproutCareTaskSchedule.swift
@@ -14,17 +14,21 @@ struct SproutCareTaskSchedule {
     let recurrenceRule: SproutCareTaskRecurrenceRule?
 
     init?(startDate: Date, dueDate: Date) {
-        guard startDate < dueDate else { return nil }
         self.init(startDate: startDate, dueDate: dueDate, recurrenceRule: nil)
     }
 
     init?(startDate: Date, recurrenceRule: SproutCareTaskRecurrenceRule) {
-        guard let nextPossibleDate = recurrenceRule.nextDate(after: Date()) else { return nil }
-
+        guard let nextPossibleDate = recurrenceRule.nextDate(after: startDate) else { return nil }
         self.init(startDate: startDate, dueDate: nextPossibleDate, recurrenceRule: recurrenceRule)
     }
 
-    init(startDate: Date, dueDate: Date, recurrenceRule: SproutCareTaskRecurrenceRule?) {
+    init?(startDate: Date, dueDate: Date, recurrenceRule: SproutCareTaskRecurrenceRule?) {
+        if recurrenceRule != nil {
+            guard dueDate == recurrenceRule?.nextDate(after: startDate) else { return nil }
+        } else {
+            guard dueDate > startDate else { return nil }
+        }
+
         self.startDate = startDate
         self.dueDate = dueDate
         self.recurrenceRule = recurrenceRule
@@ -35,3 +39,5 @@ struct SproutCareTaskSchedule {
         return formatter.string(from: self)
     }
 }
+
+extension SproutCareTaskSchedule: Equatable { }
