+++ SproutKit/Sources/SproutKit/Models/Templates/SproutPlantTemplate.swift
