--- SproutKit/Sources/SproutKit/Static Data/SproutCareTypes.swift
@@ -1,12 +1,0 @@
-//
-//  SproutCareTypes.swift
-//  Sprout
-//
-//  Created by Ryan Thally on 6/17/21.
-//
-
-import UIKit
-
-public enum SproutCareType: String, Hashable, CaseIterable {
-    case watering
-}
